backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
celery==5.3.4
aiofiles==23.2.1
pillow==10.1.0
openai==1.3.7
stability-sdk==0.8.4
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1 